[Feat] API URL 변경 사항, 클래스 라벨링 추가
- 깃 충돌 방지를 위해 라벨링을 부여하기로 함
- 판매자 관련 API URL이 변경됨에 따라 변경 사항 적용
